Introduce ImmutableAssignment to pass to Grounder
