+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/BridgedGrounder.java
@@ -1,10 +1,10 @@
 package at.ac.tuwien.kr.alpha.grounder;
 
+import at.ac.tuwien.kr.alpha.common.ImmutableAssignment;
 import at.ac.tuwien.kr.alpha.common.NoGood;
 import at.ac.tuwien.kr.alpha.common.Predicate;
 import at.ac.tuwien.kr.alpha.grounder.bridges.Bridge;
 import at.ac.tuwien.kr.alpha.grounder.parser.ParsedProgram;
-import at.ac.tuwien.kr.alpha.solver.Assignment;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -22,7 +22,7 @@ public abstract class BridgedGrounder extends AbstractGrounder {
 		this.bridges = bridges;
 	}
 
-	protected Set<NoGood> collectExternal(Assignment assignment, AtomStore atomStore) {
+	protected Set<NoGood> collectExternal(ImmutableAssignment assignment, AtomStore atomStore) {
 		Set<NoGood> collectedNoGoods = new HashSet<>();
 
 		for (Bridge bridge : bridges) {
