+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/bridges/Bridge.java
@@ -1,11 +1,11 @@
 package at.ac.tuwien.kr.alpha.grounder.bridges;
 
+import at.ac.tuwien.kr.alpha.common.ImmutableAssignment;
 import at.ac.tuwien.kr.alpha.common.NoGood;
 import at.ac.tuwien.kr.alpha.grounder.AtomStore;
-import at.ac.tuwien.kr.alpha.solver.Assignment;
 
 import java.util.Collection;
 
 public interface Bridge {
-	Collection<NoGood> getNoGoods(Assignment assignment, AtomStore atomStore);
+	Collection<NoGood> getNoGoods(ImmutableAssignment assignment, AtomStore atomStore);
 }
