+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/bridges/HexBridge.java
@@ -3,7 +3,6 @@ package at.ac.tuwien.kr.alpha.grounder.bridges;
 import at.ac.tuwien.kr.alpha.common.*;
 import at.ac.tuwien.kr.alpha.grounder.AtomId;
 import at.ac.tuwien.kr.alpha.grounder.AtomStore;
-import at.ac.tuwien.kr.alpha.solver.Assignment;
 
 import java.util.*;
 
@@ -11,7 +10,7 @@ public class HexBridge implements Bridge {
 	public static native void sendResults(String[][] resultsArray);
 	private static native String[][] externalAtomsQuery(String[] trueAtoms, String[] falseAtoms);
 
-	public Collection<NoGood> getNoGoods(Assignment assignment, AtomStore atomStore) {
+	public Collection<NoGood> getNoGoods(ImmutableAssignment assignment, AtomStore atomStore) {
 		Set<NoGood> nogoods = new HashSet<>();
 
 		List<String> trueAtoms = new ArrayList<>();
